+++ src/main/java/utils/validator/Foo.java
@@ -1,0 +1,8 @@
+package utils.validator;
+
+public class Foo implements Validator {
+	public boolean run(StringBuilder input) {
+		return true;
+	}
+}
+
